requirements.txt
Flask==0.10.1
Jinja2==2.7.3
Markdown==2.5.1
MarkupSafe==0.23
Werkzeug==0.9.6
argparse==1.2.1
itsdangerous==0.24
linecache2==1.0.0
py-gfm==0.1
six==1.9.0
traceback2==1.4.0
unittest2==1.0.1
virtualenv==12.0.2
virtualenvwrapper==4.3.1
wsgiref==0.1.2
